4-09-2023

+++ src/Formula2.java
@@ -10,9 +10,9 @@ public class Formula2 {
         short a,b;
 
         System.out.print("Enter the value (A) : ");
-        a = input.nextByte();
+        a = input.nextShort();
         System.out.print("Enter the value (B) : ");
-        b = input.nextByte();
+        b = input.nextShort();
 
         int c = (a * a) - (2*a*b) + (b * b);
